added a condition

+++ scripts/stop_container.sh
@@ -3,5 +3,8 @@ set -e
 
 # Stop the running container (if any)
 containerId=$(docker ps | awk -F " " '{print $1}')
-docker stop "$containerId"
-docker rm -f "$containerId"
+
+if [ -n "$containerId" ]; then
+  docker stop "$containerId"
+  docker rm -f "$containerId"
+fi
